+++ media-sound/alsa-driver/alsa-driver-1.0.8.ebuild
@@ -1,6 +1,6 @@
 # Copyright 1999-2005 Gentoo Foundation
 # Distributed under the terms of the GNU General Public License v2
-# $Header: /var/cvsroot/gentoo-x86/media-sound/alsa-driver/alsa-driver-1.0.8.ebuild,v 1.15 2005/06/12 17:02:33 kloeri Exp $
+# $Header: /var/cvsroot/gentoo-x86/media-sound/alsa-driver/alsa-driver-1.0.8.ebuild,v 1.16 2005/08/25 22:17:04 chriswhite Exp $
 
 IUSE="oss doc"
 inherit linux-mod flag-o-matic eutils
@@ -90,9 +90,9 @@ src_compile() {
 	# linux-mod_src_compile doesn't work well with alsa
 
 	# -j1 : see bug #71028
-	set_arch_to_kernel
+	ARCH=$(tc-arch-kernel)
 	emake -j1 || die "Make Failed"
-	set_arch_to_portage
+	ARCH=$(tc-arch)
 
 	if use doc;
 	then
@@ -158,7 +158,7 @@ pkg_postinst() {
 	einfo "Check out the ALSA installation guide availible at the following URL:"
 	einfo "http://www.gentoo.org/doc/en/alsa-guide.xml"
 
-	if kernel_is 2 6; then
+	if kernel_is 2 6 && [ -e ${ROOT}/lib/modules/${KV_FULL}/kernel/sound  ]; then
 		# Cleanup if they had older alsa installed
 		for file in $(find ${ROOT}/lib/modules/${KV_FULL}/${PN} -type f); do
 			rm -f ${file//${KV_FULL}\/${PN}/${KV_FULL}\/kernel\/sound}
